feat(csharp/src/Drivers/Apache/Spark): poc - Support for Apache Spark over HTTP (non-Arrow) (#2018)
Proof-of-concept

Adds support for standard Apache Spark data source
* Requires `adbc.spark.type` parameter option to specify the Spark
server type.
* Handles the possibility that data source does not support "catalog" in
the table object namespace (i.e., only schema and table names).

Closes #2015

+++ csharp/test/Drivers/Apache/Spark/NumericValueTests.cs
@@ -15,6 +15,7 @@
 * limitations under the License.
 */
 
+using System;
 using System.Data.SqlTypes;
 using System.Threading.Tasks;
 using Apache.Arrow.Adbc.Drivers.Apache.Hive2;
@@ -26,12 +27,12 @@ namespace Apache.Arrow.Adbc.Tests.Drivers.Apache.Spark
     // TODO: When supported, use prepared statements instead of SQL string literals
     //      Which will better test how the driver handles values sent/received
 
-    public class NumericValueTests : SparkTestBase
+    public class NumericValueTests : TestBase<SparkTestConfiguration, SparkTestEnvironment>
     {
         /// <summary>
         /// Validates that specific numeric values can be inserted, retrieved and targeted correctly
         /// </summary>
-        public NumericValueTests(ITestOutputHelper output) : base(output) { }
+        public NumericValueTests(ITestOutputHelper output) : base(output, new SparkTestEnvironment.Factory()) { }
 
         /// <summary>
         /// Validates if driver can send and receive specific Integer values correctly
@@ -112,7 +113,8 @@ namespace Apache.Arrow.Adbc.Tests.Drivers.Apache.Spark
         {
             string columnName = "SMALLNUMBER";
             using TemporaryTable table = await NewTemporaryTableAsync(Statement, string.Format("{0} DECIMAL(2,0)", columnName));
-            await ValidateInsertSelectDeleteSingleValueAsync(table.TableName, columnName, SqlDecimal.Parse(value));
+            object? expectedValue = TestEnvironment.GetValueForProtocolVersion(value, new SqlDecimal(double.Parse(value)));
+            await ValidateInsertSelectDeleteSingleValueAsync(table.TableName, columnName, expectedValue);
         }
 
         /// <summary>
@@ -127,23 +129,26 @@ namespace Apache.Arrow.Adbc.Tests.Drivers.Apache.Spark
         {
             string columnName = "SMALLNUMBER";
             using TemporaryTable table = await NewTemporaryTableAsync(Statement, string.Format("{0} DECIMAL(2,0)", columnName));
-            await Assert.ThrowsAsync<HiveServer2Exception>(async () => await ValidateInsertSelectDeleteSingleValueAsync(table.TableName, columnName, new SqlDecimal(value)));
+            await Assert.ThrowsAsync<HiveServer2Exception>(
+                async () => await ValidateInsertSelectDeleteSingleValueAsync(
+                    table.TableName,
+                    columnName, TestEnvironment.GetValueForProtocolVersion(value.ToString(), new SqlDecimal(value))));
         }
 
         /// <summary>
         /// Validates if driver can handle a large scale Number type correctly
         /// </summary>
         [SkippableTheory]
-        [InlineData("0")]
-        [InlineData("-2.003")]
-        [InlineData("4.85")]
-        [InlineData("0.0000000000000000000000000000000000001")]
+        [InlineData("0E-37")]
+        [InlineData("-2.0030000000000000000000000000000000000")]
+        [InlineData("4.8500000000000000000000000000000000000")]
+        [InlineData("1E-37")]
         [InlineData("9.5545204502636499875576383003668916798")]
         public async Task TestLargeScaleNumberRange(string value)
         {
             string columnName = "LARGESCALENUMBER";
             using TemporaryTable table = await NewTemporaryTableAsync(Statement, string.Format("{0} DECIMAL(38,37)", columnName));
-            await ValidateInsertSelectDeleteSingleValueAsync(table.TableName, columnName, SqlDecimal.Parse(value));
+            await ValidateInsertSelectDeleteSingleValueAsync(table.TableName, columnName, TestEnvironment.GetValueForProtocolVersion(value, new SqlDecimal(double.Parse(value))));
         }
 
         /// <summary>
@@ -165,7 +170,7 @@ namespace Apache.Arrow.Adbc.Tests.Drivers.Apache.Spark
         /// Validates if driver can handle a small scale Number type correctly
         /// </summary>
         [SkippableTheory]
-        [InlineData("0")]
+        [InlineData("0.00")]
         [InlineData("4.85")]
         [InlineData("-999999999999999999999999999999999999.99")]
         [InlineData("999999999999999999999999999999999999.99")]
@@ -173,7 +178,7 @@ namespace Apache.Arrow.Adbc.Tests.Drivers.Apache.Spark
         {
             string columnName = "SMALLSCALENUMBER";
             using TemporaryTable table = await NewTemporaryTableAsync(Statement, string.Format("{0} DECIMAL(38,2)", columnName));
-            await ValidateInsertSelectDeleteSingleValueAsync(table.TableName, columnName, SqlDecimal.Parse(value));
+            await ValidateInsertSelectDeleteSingleValueAsync(table.TableName, columnName, TestEnvironment.GetValueForProtocolVersion(value, SqlDecimal.Parse(value)));
         }
 
         /// <summary>
@@ -204,9 +209,9 @@ namespace Apache.Arrow.Adbc.Tests.Drivers.Apache.Spark
             SqlDecimal value = new SqlDecimal(input);
             SqlDecimal returned = new SqlDecimal(output);
             await InsertSingleValueAsync(table.TableName, columnName, value.ToString());
-            await SelectAndValidateValuesAsync(table.TableName, columnName, returned, 1);
+            await SelectAndValidateValuesAsync(table.TableName, columnName, TestEnvironment.GetValueForProtocolVersion(output.ToString(), returned), 1);
             string whereClause = GetWhereClause(columnName, returned);
-            await DeleteFromTableAsync(table.TableName, whereClause, 1);
+            if (SupportsDelete) await DeleteFromTableAsync(table.TableName, whereClause, 1);
         }
 
         /// <summary>
@@ -230,7 +235,7 @@ namespace Apache.Arrow.Adbc.Tests.Drivers.Apache.Spark
             await InsertSingleValueAsync(table.TableName, columnName, valueString);
             await SelectAndValidateValuesAsync(table.TableName, columnName, value, 1);
             string whereClause = GetWhereClause(columnName, value);
-            await DeleteFromTableAsync(table.TableName, whereClause, 1);
+            if (SupportsDelete) await DeleteFromTableAsync(table.TableName, whereClause, 1);
         }
 
         /// <summary>
@@ -255,9 +260,11 @@ namespace Apache.Arrow.Adbc.Tests.Drivers.Apache.Spark
             using TemporaryTable table = await NewTemporaryTableAsync(Statement, string.Format("{0} FLOAT", columnName));
             string valueString = ConvertFloatToString(value);
             await InsertSingleValueAsync(table.TableName, columnName, valueString);
-            await SelectAndValidateValuesAsync(table.TableName, columnName, value, 1);
+            object doubleValue = (double)value;
+            object floatValue = TestEnvironment.GetValueForProtocolVersion(doubleValue, value)!;
+            await base.SelectAndValidateValuesAsync(table.TableName, columnName, floatValue, 1);
             string whereClause = GetWhereClause(columnName, value);
-            await DeleteFromTableAsync(table.TableName, whereClause, 1);
+            if (SupportsDelete) await DeleteFromTableAsync(table.TableName, whereClause, 1);
         }
     }
 }
